Add vertex descriptors

+++ game/src/level.rs
@@ -527,9 +527,9 @@ impl<'a> Builder<'a> {
             a_tile_uv: [tile_u, tile_v],
             a_tile_size: [bounds.size[0], bounds.size[1]],
             a_scroll_rate: scroll_rate,
-            a_num_frames: bounds.num_frames as u8,
+            a_num_frames: bounds.num_frames as u32,
             a_row_height: bounds.row_height as f32,
-            a_light: light_info,
+            a_light: light_info as u32,
             ..Default::default()
         });
         self
@@ -542,9 +542,9 @@ impl<'a> Builder<'a> {
             a_tile_uv: [-xz[0] * 100.0, -xz[1] * 100.0],
             a_tile_size: [bounds.size[0], bounds.size[1]],
             a_scroll_rate: 0.0,
-            a_num_frames: bounds.num_frames as u8,
+            a_num_frames: bounds.num_frames as u32,
             a_row_height: bounds.row_height as f32,
-            a_light: light_info,
+            a_light: light_info as u32,
             ..Default::default()
         });
         self
@@ -574,7 +574,7 @@ impl<'a> Builder<'a> {
             a_tile_uv: [tile_u, tile_v],
             a_tile_size: [bounds.size[0], bounds.size[1]],
             a_num_frames: 1,
-            a_light: light_info,
+            a_light: light_info as u32,
             ..Default::default()
         });
         self
